requirements.txt
asgiref==3.8.1
attrs==24.2.0
black==24.8.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.6.1
distlib==0.3.8
Django==5.1.1
django-admin-thumbnails==0.2.9
django-js-asset==2.2.0
django-mptt==0.16.0
django-session-timeout==0.1.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
filelock==3.16.0
flake8==7.1.1
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==2.0.1
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.2
pluggy==1.5.0
pycodestyle==2.12.1
pyflakes==3.2.0
pytest==8.3.3
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
virtualenv==20.26.4
